adjust filepath so tests can run
