+++ samples/interactive-tutorials/search/search_simple_query_test.py
@@ -22,7 +22,7 @@ from search_simple_query import search
 
 def test_search_simple_query_pass():
     output = str(
-        subprocess.check_output("python search/search_simple_query.py", shell=True)
+        subprocess.check_output("python search_simple_query.py", shell=True)
     )
 
     assert re.match(".*search request.*", output)
