+++ samples/interactive-tutorials/search/search_with_ordering_test.py
@@ -22,7 +22,7 @@ from search_with_ordering import search
 
 def test_search_with_ordering_pass():
     output = str(
-        subprocess.check_output("python search/search_with_ordering.py", shell=True)
+        subprocess.check_output("python search_with_ordering.py", shell=True)
     )
 
     assert re.match(".*search request.*", output)
